move add entity menu to scene graph
